docshell/test/mochitest/test_bug1045096.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1045096
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1045096</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1045096">Mozilla Bug 1045096</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 1045096 **/
  var i = document.createElement("iframe");
  i.src = "javascript:false"; // This is required!
  $("content").appendChild(i);
  ok(i.contentWindow.performance, "Should have a performance object");
  </script>
</body>
</html>
